Bug 49619 - [formatting] comment formatter leaves whitespace in comments
With this, the Javadoc formatter preserves a trailing space if it was
there already, but does not add a new one. Also adjusted the tests which
expected that a new trailing space was added.

Note that the block comment formatter is not fixed by this, it still has
the same problem.

Change-Id: I157a12661d5203c4a983b89ae7531d5b71b29bcd
Signed-off-by: Robin Stocker <robin@nibor.org>