Bug 441726 - JDT performance regression due to bug 410207
Add a cache in JavaModelManager that maps a jar's IPath to its Java
language level, and persist the cache across Eclipse sessions. This
cache makes JarPackageFragmentRoot creation much faster. Items in the
cache are invalidated whenever the DeltaProcessor sees that a jar's
timestamp has changed.

FullSourceWorkspaceModelTests.testGetAllPackageFragmentRoots took 123ms
to execute before this patch, and only 29ms with this patch applied.
The same test took 30ms in Kepler with the patch for bug 411423 applied.

Change-Id: Ic97dff022f9501a25315c35fae42df79b01486cc
Signed-off-by: Terry Parker <tparker@google.com>