Bug 514653 - JobManager sleep time is longer then indexing itself
Only sleep after submitting a job if job indicates that this is really
needed. Also reduce wait time from 50 to 10 ms, which seem to be more
appropriate with today's hardware.

The main part of the speedup is coming from the IndexRequest instances
scheduled via IndexManager.scheduleDocumentIndexing(...), which do NOT
require an implicit wait now because they are doing only IO read and
in-memory index write operations.

Change-Id: I8b52405accdb643dd494a73979accd41b8d9f5b7
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>