Bug 531382 - Batch compile ignores source from build participant
Whenever the limit "maxCompiledUnitsAtOnce" (VM argument for Eclipse) is
exceeded, sources generated by build participants are not compiled.

This is the case, since AbstractImageBuilder.compile will check its work
queue before compiling in batch mode. The queue however contains only
the sources with which the build was triggered. It does not contain
sources generated by build participants. So such generated sources are
ignored during the batch compile.

With this change, the work queue of the builder is updated to contain
also sources from the build participants. The correct behavior is
validated by a regression test.

Change-Id: Ia1b0598bbacefa7c10034f510e59f31b5c225a9e
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>