Bug 413114 - added encoding assertions in test testBug376673e
If JavaSearchBugsTests2.testBug376673e fails in Linux environments if
the environment variable LANG is not set to UTF-8. This is e.g. the case
in official builds.

This change adds further assertions in the test, to validate that
correct encoding is set. If the encoding is not as expected, the test
fails with a more comprehensible message.

Change-Id: I21fe15947262137f48195026add56525fe35a748
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>