Bug 564905 - [regression] Missing full build on .class file changes
If the output folder of a project is deleted in file system, while the
project is closed, the project is not built on re-open/refresh. The same
is true if Eclipse was closed when deleting the output folder (e.g. with
a git clean).

This is a regression caused by a fix for bug 563030. When
State.typeLocators is persisted, instead of writing key/value pairs,
value/value pairs are written. As a result, after persisting the project
build state, IncrementalImageBuilder.checkForClassFileChanges() no
longer detects .class file changes. It looks with keys in
State.typeLocators, while the map has value->value pairs.

Change-Id: I154764626d7a6ff264c0d3d4931d1689df60878e
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>