Bug 563628 - NullPointerException in org.eclipse.jdt.internal.compiler.tool.EclipseFileManager.getLocationForModule
In this specific case the module is different from project folder and
the java files are passed to to the compiler without modulesourcepath.
In this case we ger a null when we try to find output location for a
module.

The solution is to create modules in the standard CLASS_OUTPUT folder.

Change-Id: I09da1602ce5559fdd7678c491fc992d6f0661c91
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>