Bug 571258 - [test] randomly failing AllJavaModelTestsTracing
If we assume, that sometimes delta processing is done in a asynchronous
way, creating a project can have indeterminstic result in a delta or not
wether or not the asynchronous job is in time or not.

Let try to move "startDeltas" before creating the project, so we capture
and expect project creation delta.

Additionally add delta information to the delta trace.

Change-Id: Iad12c01f6fd40b360039b93c90056b07cbf464d4
Signed-off-by: jkubitz <jkubitz-eclipse@gmx.de>