Bug 572213 - Dogfooding - Use "Exit loop earlier" in JDT core
Changes the loop in Javadoc to be finished by a break statement. This
way it is handled as the rest of the JDT core, using a break statement
instead of an additional condition. This aligns the usage of a break
pattern in the JDT core code base instead of using a different pattern
to exit a loop in one place.

Change-Id: I4a7223a053d11d7e68cc2c6d88a370b0ebb74619
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>